Projekt_Docker/requirements.txt
mysql-connector-python==8.0.28
Faker==20.1.0
aiomysql==0.0.22
